Backend/requirements.txt
django
pyjwt
djangorestframework
qrcode